[inert] Let modal dialogs escape inertness from ancestors
If an ancestor of a modal <dialog> has the 'inert' attribute, the dialog
and its subtree will no longer be inert.

However, if the 'inert' attribute is in the <dialog> itself, then the
subtree will still be inert.

This aligns Blink with Gecko and WebKit. The spec conversation about
this topic is in https://github.com/whatwg/html/issues/7808

Bug: 1323680

TEST=external/wpt/inert/inert-with-modal-dialog-001.html
TEST=StyleResolverTest.IsInertWithAttributeAndDialog

Change-Id: Id2752ce3d1c8985b68b1d700710c3ff3eb100714