Remove ClampTo() on CreateArithmeticOperationSimplified()
CreateArithmeticOperationSimplified() calculates the simplification of math expression on the parse time. But the internal values of css math expression must handle the degenerated values (e.g. infinity and NaN).

And the values are clamped when the other components consume them.

Therefore, remove ClampTo() for parsing the additive expression of CreateArithmeticOperationSimplified().

Bug: 1313292
Change-Id: Ie98b764e315df58bc37a4de9523fb6f746f7181d